+++ llarp/dht/localserviceaddresslookup.hpp
@@ -14,7 +14,7 @@ namespace llarp
       PathID_t localPath;
 
       LocalServiceAddressLookup(const PathID_t &pathid, uint64_t txid,
-                                uint32_t relayOrder, const Key_t &addr,
+                                uint64_t relayOrder, const Key_t &addr,
                                 AbstractContext *ctx,
                                 __attribute__((unused)) const Key_t &askpeer);
