+++ llarp/dht/messages/findintro.hpp
@@ -18,9 +18,9 @@ namespace llarp
       llarp::service::Tag tagName;
       uint64_t txID       = 0;
       bool relayed        = false;
-      uint32_t relayOrder = 0;
+      uint64_t relayOrder = 0;
 
-      FindIntroMessage(const Key_t& from, bool relay, uint32_t order)
+      FindIntroMessage(const Key_t& from, bool relay, uint64_t order)
           : IMessage(from)
       {
         relayed    = relay;
@@ -38,7 +38,7 @@ namespace llarp
       }
 
       explicit FindIntroMessage(uint64_t txid, const Key_t& addr,
-                                uint64_t maxRecursionDepth, uint32_t order)
+                                uint64_t maxRecursionDepth, uint64_t order)
           : IMessage({})
           , recursionDepth(maxRecursionDepth)
           , location(addr)
